refactor: move OAuth related classes into oauth2 package
