+++ src/main/java/eu/bbmri_eric/negotiator/common/configuration/security/UserPrincipal.java
@@ -1,5 +1,7 @@
-package eu.bbmri_eric.negotiator.user;
+package eu.bbmri_eric.negotiator.common.configuration.security;
 
+import eu.bbmri_eric.negotiator.user.NegotiatorPrincipal;
+import eu.bbmri_eric.negotiator.user.Person;
 import lombok.Getter;
 
 @Getter
