+++ src/main/java/eu/bbmri_eric/negotiator/common/configuration/security/oauth2/CustomJWTAuthConverter.java
@@ -1,4 +1,4 @@
-package eu.bbmri_eric.negotiator.common.configuration.security.auth;
+package eu.bbmri_eric.negotiator.common.configuration.security.oauth2;
 
 import eu.bbmri_eric.negotiator.common.exceptions.WrongJWTException;
 import eu.bbmri_eric.negotiator.user.Person;
