+++ src/test/java/eu/bbmri_eric/negotiator/config/WithMockCustomUserSecurityContextFactory.java
@@ -1,6 +1,6 @@
 package eu.bbmri_eric.negotiator.config;
 
-import eu.bbmri_eric.negotiator.common.configuration.security.auth.NegotiatorJwtAuthenticationToken;
+import eu.bbmri_eric.negotiator.common.configuration.security.oauth2.NegotiatorJwtAuthenticationToken;
 import eu.bbmri_eric.negotiator.user.Person;
 import eu.bbmri_eric.negotiator.util.WithMockNegotiatorUser;
 import java.time.Instant;
